# openapi.yaml
openapi: 3.0.0
info:
  title: Online Store API
  description: API documentation for managing users, products, and user roles in an online store.
  version: 1.0.0

servers:
  - url: http://localhost:8081/api
    description: Local development server

paths:
  /users:
    post:
      summary: Create or update a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
        '500':
          description: Internal server error

  /users/{id}:
    get:
      summary: Retrieve a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    delete:
      summary: Delete a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
      responses:
        '204':
          description: User deleted successfully
        '500':
          description: Internal server error

  /users/{uid}/role/admin:
    post:
      summary: Assign admin role to a user
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
          description: Firebase UID of the user
      responses:
        '200':
          description: Admin role assigned successfully
        '500':
          description: Internal server error

  /users/{uid}/role/customer:
    post:
      summary: Assign customer role to a user
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
          description: Firebase UID of the user
      responses:
        '200':
          description: Customer role assigned successfully
        '500':
          description: Internal server error

  /products:
    get:
      summary: Retrieve a list of all products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created successfully
        '500':
          description: Internal server error

  /products/{id}:
    get:
      summary: Retrieve a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
    delete:
      summary: Delete a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '204':
          description: Product deleted successfully
        '500':
          description: Internal server error

  /categories:
    get:
      summary: Retrieve a list of all categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '201':
          description: Category created successfully
        '500':
          description: Internal server error

  /categories/{id}:
    get:
      summary: Retrieve a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Category ID
      responses:
        '200':
          description: Category retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
    delete:
      summary: Delete a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Category ID
      responses:
        '204':
          description: Category deleted successfully
        '500':
          description: Internal server error

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        address:
          type: string

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
          format: int32

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string